backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
python-dotenv==1.0.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
PyJWT==2.8.0
uritemplate==4.1.1
coreapi==2.3.3
coreschema==0.0.4
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
opencv-python-headless==4.9.0.80
python-ulid==2.2.0
django-cors-headers==3.14.0
